pyproject.toml
[tool.poetry]
name = "class2019bot"
version = "0.1.0"
description = "The telegram bot for the text adventure games made in the comp app chem class of 2019."
authors = ["Class of 2019, Chulalongkorn University, BSAC Chemistry"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^13.0"
PyYAML = "^5.3.1"
colorlog = "^4.4.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py37']

